dashboard/requirements.txt
streamlit==1.37.1
kafka-python==2.0.2
pandas==2.2.2
